backend/requirements.txt
Django==5.2.1
djangorestframework
channels
channels-redis
django-cors-headers
daphne
redis
Unidecode
python-dateutil
Pillow

